LayoutTests/platform/ios/http/tests/site-isolation/pointer-lock-expected.txt

CONSOLE MESSAGE: Pointerlock not supported
